Move player_stats function into save_data
Update_player moved into the save_data script
Change other scripts to use the function from this script
