+++ hangman_utils/__init__.py
@@ -1,9 +1,7 @@
 from hangman_utils.hangman import HangingMan
-from hangman_utils.save_data import load_game_data
-from hangman_utils.player_stats import load_data
+from hangman_utils.save_data import load_data
 
 __all__ = [
     'HangingMan',
-    'load_game_data',
     'load_data',
 ]
