+++ hangman_utils/hangman.py
@@ -2,8 +2,7 @@ import os
 import random
 
 from .life_bars import LIFE
-from .save_data import update_match
-from .player_stats import update_player
+from .save_data import update_match, update_player
 
 
 ALPHABET = 'abcdefghijklmnopqrstuvwxyz'
